add rework commit and product receipt

+++ aaschina/aas_mes/static/src/js/mes_gp12_checking.js
@@ -4,6 +4,13 @@
 
 $(function() {
 
+    $('input[type="radio"].flat-blue').iCheck({ checkboxClass: 'icheckbox_flat-blue', radioClass: 'iradio_flat-blue'});
+
+    $('input[type="radio"].flat-blue').on('ifChecked', function(event){
+        var labelqty = $(this).attr('qty');
+        $('#mes_labelqty').attr('qty', labelqty);
+    });
+
     var scanable = true; //是否可以继续扫描
     new VScanner(function(barcode) {
         if (barcode == null || barcode == '') {
@@ -104,10 +111,14 @@ $(function() {
                 $('#check_result').html(dresult.result);
                 if(dresult.result=='OK'){
                     $('#check_result_box').removeClass('bg-red').addClass('bg-green');
-                    $('#pass_list').append('<tr><td>'+dresult.operate_result+'</td></tr>');
+                    var serialtr = $('<tr></tr>').prependTo($('#pass_list')).html('<td>'+dresult.operate_result+'</td>');
+                    serialtr.attr({'id': 'serialnumber_'+dresult.serialnumber_id, 'serialnumberid': dresult.serialnumber_id});
+                    serialtr.addClass('aas-waitprint');
                     var scount = parseInt($('#pass_count').attr('scount'));
                     scount += 1;
                     $('#pass_count').attr('scount', scount).html(scount);
+                    var waitcount = parseInt($('#mes_printbtn').attr('waitcount'));
+                    $('#mes_printbtn').attr('waitcount', waitcount+1);
                 }else{
                     $('#check_result_box').removeClass('bg-green').addClass('bg-red');
                     $('#fail_list').append('<tr><td>'+dresult.operate_result+'</td></tr>');
@@ -150,6 +161,11 @@ $(function() {
                     });
                 }
                 $('#mes_serialnumber').html(barcode);
+                var waitcount = parseInt($('#mes_printbtn').attr('waitcount'));
+                var labelqty = parseInt($('#mes_labelqty').attr('qty'));
+                if(waitcount >= labelqty){
+                    action_print_label();
+                }
             },
             error:function(xhr,type,errorThrown){
                 scanable = true;
@@ -157,6 +173,70 @@ $(function() {
             }
         });
     }
+
+    //自动生成标签并打印
+    function action_print_label(){
+        var serialnumberlist = $('.aas-waitprint');
+        if(serialnumberlist.length <= 0){
+            layer.msg('没有成品可以打印标签！', {icon: 5});
+            return ;
+        }
+        var serialnumberids = [];
+        $.each(serialnumberlist, function(index, serialnumbertr){
+            serialnumberids.push(parseInt(serialnumbertr.attr('serialnumberid')));
+        });
+        var printerid = parseInt($('#mes_printer').val());
+        if(printerid==0){
+            layer.msg('您还未选择标签打印机，请先设置好标签打印机！', {icon: 5});
+            return ;
+        }
+        var scanparams = {'serialnumberids': serialnumberids, 'printer_id': printerid};
+        var access_id = Math.floor(Math.random() * 1000 * 1000 * 1000);
+        $.ajax({
+            url: '/aasmes/gp12/dolabel',
+            headers:{'Content-Type':'application/json'},
+            type: 'post', timeout:10000, dataType: 'json',
+            data: JSON.stringify({ jsonrpc: "2.0", method: 'call', params: scanparams, id: access_id}),
+            success:function(data){
+                var dresult = data.result;
+                if(!dresult.success){
+                    layer.msg(dresult.message, {icon: 5});
+                    return ;
+                }
+                $.each(dresult.records, function(index, record){
+                    var params = {'label_name': dresult.printer, 'label_count': 1, 'label_content':record};
+                    $.ajax({type:'post', dataType:'script', url:'http://'+dresult.serverurl, data: params,
+                        success: function (result) { },
+                        error:function(XMLHttpRequest,textStatus,errorThrown){}
+                    });
+                });
+                serialnumberlist.removeClass('aas-waitprint');
+                $('#mes_printbtn').attr('waitcount', 0);
+                $('#mes_label').html(dresult.label_name);
+            },
+            error:function(xhr,type,errorThrown){
+                scanable = true;
+                console.log(type);
+            }
+        });
+    }
+
+    //标签打印，可能未达到自动打印标签的数量
+    $('#mes_printbtn').click(function(){
+        var waitcount = parseInt($(this).attr('waitcount'));
+        if(waitcount==0){
+            layer.msg('当前还无法打印标签', {icon: 5});
+            return ;
+        }
+        var labelqty = parseInt($('#mes_labelqty').attr('qty'));
+        var tipmessage = '您确认要打印标签吗？';
+        if(waitcount < labelqty){
+            tipmessage = '当前成品数量少于'+labelqty+',您确认要打印标签吗？';
+        }
+        layer.confirm(tipmessage, {'btn': ['确定', '取消']}, function(){
+            action_print_label();
+        },function(){});
+    });
 
     //单击员工清单
     $('#employee_list').on('click', '.aas-employee', function(){
@@ -168,17 +248,53 @@ $(function() {
 
     //上报不良
     $('#action_badmode').click(function(){
-        var employeeid = parseInt($('#mes_operator').attr('employeeid'));
-        if(employeeid==0){
-            layer.msg('请在左侧员工列表中选择一个当前操作员工', {icon: 5});
-            return ;
+        layer.confirm('您确认上报不良？', {'btn': ['确定', '取消']}, function(){
+            var employeeid = parseInt($('#mes_operator').attr('employeeid'));
+            if(employeeid==0){
+                layer.msg('请在左侧员工列表中选择一个当前操作员工', {icon: 5});
+                return ;
+            }
+            localStorage.setItem('employeeid', employeeid);
+            var employeename = $('#employee_'+employeeid).attr('employeename');
+            if(employeename!=undefined&&employeename!=null&&employeename!=''){
+                localStorage.setItem('employeename', employeename);
+            }
+            window.location.replace('/aasmes/gp12/rework');
+        }, function(){});
+    });
+
+    //成品入库
+    $('#action_delivery').click(function(){
+        layer.confirm('您确认成品出货？', {'btn': ['确定', '取消']}, function(){
+            window.location.replace('/aasmes/gp12/delivery');
+        }, function(){});
+    });
+
+    $('#mes_printer').select2({
+        placeholder: '选择标签打印机...',
+        ajax:{
+            type: 'post',
+            timeout:10000,
+            dataType: 'json',
+            url: '/aasmes/printerlist',
+            headers:{'Content-Type':'application/json'},
+            data: function(params){
+                var sparams =  {
+                    'q': params.term || '', 'page': params.page || 1
+                };
+                return JSON.stringify({ jsonrpc: "2.0", method: 'call', params: sparams, id: Math.floor(Math.random() * 1000 * 1000 * 1000) })
+            },
+            processResults: function (data, params) {
+                params.page = params.page || 1;
+                var dresult = data.result;
+                return {
+                    results: dresult.items,
+                    pagination: {
+                        more: (params.page * 30) < dresult.total_count
+                    }
+                };
+            }
         }
-        localStorage.setItem('employeeid', employeeid);
-        var employeename = $('#employee_'+employeeid).attr('employeename');
-        if(employeename!=undefined&&employeename!=null&&employeename!=''){
-            localStorage.setItem(('employeename', employeename));
-        }
-        window.location.replace('/aasmes/gp12/rework');
     });
 
 });
